+++ Classes/Controller/UploadController.php
@@ -1,0 +1,53 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the TYPO3 extension: filetransfer.
+ *
+ * It is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU General Public License, either version 2
+ * of the License, or any later version.
+ *
+ * For the full copyright and license information, please read the
+ * LICENSE.txt file that was distributed with this source code.
+ *
+ * The TYPO3 project - inspiring people to share!
+ */
+
+namespace Wacon\Filetransfer\Controller;
+
+use Psr\Http\Message\ResponseInterface;
+use TYPO3\CMS\Extbase\Mvc\Controller\ActionController;
+
+final class UploadController extends ActionController
+{
+    /**
+     * Show the upoad form
+     * @return ResponseInterface
+     */
+    public function formAction(): ResponseInterface
+    {
+        return $this->htmlResponse();
+    }
+
+    /**
+     * Show the upoad form
+     * @param Upload $upload
+     * @return ResponseInterface
+     */
+    public function uploadAction(Upload $upload): ResponseInterface
+    {
+        debug($upload);
+        return $this->htmlResponse();
+    }
+
+    /**
+     * Show the success message
+     * @return ResponseInterface
+     */
+    public function successAction(): ResponseInterface
+    {
+        return $this->htmlResponse();
+    }
+}