+++ CHANGES.md
@@ -1,3 +1,8 @@
+## Version 0.0.2
+- [FEATURE] Create Upload Model
+- [FEATURE] Register Typoscript
+- [FEATURE] Register Upload Plugin with controller and view
+
 ## Version 0.0.1
 - [IMPORTANT] Create github repository
 - [IMPORTANT] Kickstart Extension